move env editor to folder tab
