+++ packages/insomnia/src/ui/routes/modals.tsx
@@ -7,7 +7,6 @@ import { AddKeyCombinationModal } from '../components/modals/add-key-combination
 import { AlertModal } from '../components/modals/alert-modal';
 import { AskModal } from '../components/modals/ask-modal';
 import { CodePromptModal } from '../components/modals/code-prompt-modal';
-import { EnvironmentEditModal } from '../components/modals/environment-edit-modal';
 import { ErrorModal } from '../components/modals/error-modal';
 import { FilterHelpModal } from '../components/modals/filter-help-modal';
 import { GenerateCodeModal } from '../components/modals/generate-code-modal';
@@ -70,10 +69,6 @@ const Modals: FC = () => {
         />
         <ResponseDebugModal
           ref={instance => registerModal(instance, 'ResponseDebugModal')}
-        />
-
-        <EnvironmentEditModal
-          ref={instance => registerModal(instance, 'EnvironmentEditModal')}
         />
 
         <AddKeyCombinationModal
